Add files via upload
Overview:-
This repository contains a simple Amazon clone built using HTML and CSS. It serves as a beginner-friendly project to practice front-end development skills. The project replicates the core features and layout of the Amazon website, providing a hands-on experience in structuring and styling web pages.

Features:-
1.)  Responsive design for a seamless user experience on various devices.
2.)  Product listings with images, titles, prices, and ratings.
3.)  Interactive elements, such as hover effects and button interactions.
4.)  Clean and modular code structure for easy understanding and customization.

How to Contribute:-
If you'd like to contribute to the project, feel free to submit pull requests. Here are some areas where you can contribute:
1.)  Improve responsiveness for additional screen sizes.
2.)  Add new product categories or features.
3.)  Enhance the overall design and aesthetics.

Acknowledgments:-
This project was inspired by the Amazon website and serves as a learning resource for web development enthusiasts.

License:-
This project is licensed under the MIT License - see the LICENSE file for details.

Author:-
Anubhav Pandey